DX-981 Make auth login an api
